requirements.txt
streamlit==1.12.0
xmltodict==0.13.0
PyGithub==1.55
johnnydep==1.11
lxml==4.9.0
pandas==1.3.5
beautifulsoup4==4.11.1
dicttoxml==1.7.4